refactor: 페이지별 모바일 헤더 UI 분기처리를 위해 검색UI 컴포넌트로 묶기 - MobileSearchWrap 컴포넌트 (#58)
