+++ src/components/layout/mobile/index.ts
@@ -1,3 +1,4 @@
 export { MobileHeader } from './MobileHeader';
 export { MobileNavBar } from './MobileNavBar';
 export { SearchHistoryDropdown } from './SearchHistoryDropdown';
+export { default as MobileSearchWrap } from './MobileSearchWrap';
